requirements.txt
pyTelegramBotAPI==4.27.0
Flask==3.1.1
Flask-CORS==6.0.1
python-dotenv==1.1.0
requests==2.32.4
schedule==1.2.2

